fix: remove unused join import and extract archive routes to fix lint
- Remove unused `import { join } from 'path'` in gitService.js (no-unused-vars error)
- Extract archive/unarchive/star routes to sessions-archive.js sub-router
  to bring sessions.js under the 500-line max-lines limit
- Add timeout to flaky test case
