+++ packages/server/src/services/gitService.js
@@ -1,8 +1,6 @@
 import { exec } from 'child_process';
 import { promisify } from 'util';
 
-import { join } from 'path';
-
 const execAsync = promisify(exec);
 
 // Cache for default branch detection per repository
